adjusted file formats and I think I finished the server.js
