--- server.js
@@ -1,33 +1,0 @@
-const express = require('express');
-const { ApolloServer } = require('apollo-server-express');
-const { typeDefs, resolvers } = require('./schema');
-const mongoose = require('mongoose');
-const cors = require('cors');
-
-const app = express();
-
-app.use(cors());
-
-mongoose.connect(process.env.MONGO_URI, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-    useFindAndModify: false    
-    .then(() => console.log('DB connected'))
-    .catch(err => console.log(err))
-});
-
-const server = new ApolloServer({
-    typeDefs,
-    resolvers,
-    context: ({ req }) => {
-        const token = req.headers.authorization || '';
-        return { token };
-    }
-});
-
-server.applyMiddleware({ app });
-
-const PORT = process.env.PORT || 3001;
-app.listen({ port: PORT }, () => {
-    console.log(`🚀 Server ready at http://localhost:${PORT}${server.graphqlPath}`);
-});
