+++ server/schema/index.js
@@ -1,0 +1,4 @@
+const typeDefs = require('./typeDefs');
+const resolvers = require('./resolvers');
+
+module.exports = { typeDefs, resolvers };
